Flink: refactor sink tests to reduce the number of combinations with parameterized tests.
E.g. it is not really necessary to test different file format for distribution mode (hash, range). I found this necessary when trying to add new tests for the new range distribution feature. We often run 20+ combinations for every test/method in some classes.
